backend/requirements.txt
fastapi
uvicorn
jinja2
langchain
langchain-core
langchain-community
langchain-ollama
python-dotenv
